+++ packages/schema/src/index.ts
@@ -5,3 +5,4 @@
 
 export * from "./entity";
 export * from "./overlay";
+export * from "./ruleset";